Removed dynamic Hardhat dependency in resolveAsset so we no longer need resolversNoHardhat.js
