+++ contracts/tasks/validator.js
@@ -11,7 +11,7 @@ const { checkPubkeyFormat } = require("./taskUtils");
 const { getValidator, getEpoch } = require("./beaconchain");
 const { storePrivateKeyToS3 } = require("../utils/amazon");
 const addresses = require("../utils/addresses");
-const { resolveContract } = require("../utils/resolversNoHardhat");
+const { resolveContract } = require("../utils/resolvers");
 const { getDefenderSigner } = require("../utils/signersNoHardhat");
 const { sleep } = require("../utils/time");
 const { logTxDetails } = require("../utils/txLogger");
