+++ contracts/utils/resolvers.js
@@ -1,7 +1,6 @@
 const addresses = require("./addresses");
 const { ethereumAddress } = require("./regex");
 const { networkMap } = require("./hardhat-helpers");
-const { resolveContract } = require("./resolversNoHardhat");
 
 const log = require("./logger")("task:assets");
 
@@ -10,15 +9,14 @@ const log = require("./logger")("task:assets");
  * @param {string} symbol token symbol of the asset. eg OUSD, USDT, stETH, CRV...
  */
 const resolveAsset = async (symbol) => {
-  // dynamically load in function so this function can be used by tasks
-  // if put outside this function, the following error occurs:
-  // "Hardhat can't be initialized while its config is being defined"
-  const hre = require("hardhat");
+  const { chainId } = await ethers.getDefaultProvider().getNetwork();
 
-  // Not using helpers here as they import hardhat which won't work for Hardhat tasks
-  if (process.env.FORK === "true" || hre.network.name != "hardhat") {
-    const { chainId } = await ethers.getDefaultProvider().getNetwork();
-    const network = networkMap[chainId] || "mainnet";
+  // If not a unit test or task running against the hardhat network
+  if (chainId != networkMap.hardhat) {
+    const network = networkMap[chainId];
+    if (!network) {
+      throw Error(`Failed to resolve network with chain Id "${chainId}"`);
+    }
 
     const assetAddr =
       addresses[network][symbol + "Proxy"] || addresses[network][symbol];
@@ -36,6 +34,47 @@ const resolveAsset = async (symbol) => {
     throw Error(`Failed to resolve symbol "${symbol}" to a mock contract`);
   }
   return asset;
+};
+
+/**
+ * Returns a contract instance.
+ * @param {string} proxy Address or name of the proxy contract or contract name if no proxy. eg OETHVaultProxy or OETHZapper
+ * @param {string} [abiName=proxy] ABI name. Will default to proxy is not used. eg VaultAdmin, VaultCore, Governable or IERC20Metadata
+ * @returns
+ */
+const resolveContract = async (proxy, abiName) => {
+  const { chainId } = await ethers.getDefaultProvider().getNetwork();
+  const networkName = networkMap[chainId];
+
+  // If proxy is an address
+  if (proxy.match(ethereumAddress)) {
+    if (!abiName) {
+      throw Error(`Must pass an ABI name if the proxy is an address`);
+    }
+    const contract = await ethers.getContractAt(abiName, proxy);
+    if (!contract) {
+      throw Error(`Failed find ABI for "${abiName}"`);
+    }
+    return contract;
+  }
+
+  const proxyContract = await ethers.getContract(proxy);
+  if (!proxyContract) {
+    throw Error(`Failed find proxy "${proxy}" on the ${networkName} network`);
+  }
+  log(
+    `Resolved proxy ${proxy} on the ${networkName} network to ${proxyContract.address}`
+  );
+
+  if (abiName) {
+    const contract = await ethers.getContractAt(abiName, proxyContract.address);
+    if (!contract) {
+      throw Error(`Failed find ABI for "${abiName}"`);
+    }
+    return contract;
+  }
+
+  return proxyContract;
 };
 
 module.exports = {
